Moving configline classes to individual files
