+++ block_filtered_course_list.php
@@ -218,7 +218,7 @@ class block_filtered_course_list extends block_base {
         $this->rubrics = array_reduce(
             array_filter(
                 array_map(function($config) {
-                    $classname = "block_filtered_course_list_{$config[0]}_configline";
+                    $classname = "\\block_filtered_course_list\\{$config[0]}_filter";
                     if (class_exists($classname)) {
                         $item = new $classname($config, $this->mycourses, $this->fclconfig);
                         return $item->get_rubrics();
